+++ src/boot/z_std_dma.c
@@ -43,7 +43,7 @@ s32 DmaMgr_CompareName(const char* name1, const char* name2) {
     return 0;
 }
 
-s32 DmaMgr_DmaRomToRam(u32 rom, u32 ram, u32 size) {
+s32 DmaMgr_DmaRomToRam(uintptr_t rom, uintptr_t ram, u32 size) {
     OSIoMesg ioMsg;
     OSMesgQueue queue;
     OSMesg msg;
@@ -164,8 +164,8 @@ void DmaMgr_DmaFromDriveRom(u32 ram, u32 rom, u32 size) {
 }
 
 void DmaMgr_Error(DmaRequest* req, const char* file, const char* errorName, const char* errorDesc) {
-    u32 vrom = req->vromAddr;
-    u32 ram = (u32)req->dramAddr;
+    uintptr_t vrom = req->vromAddr;
+    uintptr_t ram = (uintptr_t)req->dramAddr;
     u32 size = req->size;
     char buff1[80];
     char buff2[80];
@@ -196,7 +196,7 @@ void DmaMgr_Error(DmaRequest* req, const char* file, const char* errorName, cons
     Fault_AddHungupAndCrashImpl(buff1, buff2);
 }
 
-const char* DmaMgr_GetFileNameImpl(u32 vrom) {
+const char* DmaMgr_GetFileNameImpl(uintptr_t vrom) {
     DmaEntry* iter = gDmaDataTable;
     const char** name = sDmaMgrFileNames;
 
@@ -227,10 +227,10 @@ const char* DmaMgr_GetFileName(u32 vrom) {
 }
 
 void DmaMgr_ProcessMsg(DmaRequest* req) {
-    u32 vrom = req->vromAddr;
+    uintptr_t vrom = req->vromAddr;
     void* ram = req->dramAddr;
     u32 size = req->size;
-    u32 romStart;
+    uintptr_t romStart;
     u32 romSize;
     u8 found = false;
     DmaEntry* iter;
@@ -257,7 +257,7 @@ void DmaMgr_ProcessMsg(DmaRequest* req) {
                                  "セグメント境界をまたがってＤＭＡ転送することはできません");
                 }
 
-                DmaMgr_DmaRomToRam(iter->romStart + (vrom - iter->vromStart), (u32)ram, size);
+                DmaMgr_DmaRomToRam(iter->romStart + (vrom - iter->vromStart), ram, size);
                 found = true;
 
                 if (0) {
@@ -297,7 +297,7 @@ void DmaMgr_ProcessMsg(DmaRequest* req) {
             return;
         }
 
-        DmaMgr_DmaRomToRam(vrom, (u32)ram, size);
+        DmaMgr_DmaRomToRam(vrom, (uintptr_t)ram, size);
 
         if (0) {
             osSyncPrintf("No Press ROM:%08X RAM:%08X SIZE:%08X (非公式)\n", vrom, ram, size);
@@ -332,7 +332,7 @@ void DmaMgr_ThreadEntry(void* arg0) {
     osSyncPrintf("ＤＭＡマネージャスレッド実行終了\n");
 }
 
-s32 DmaMgr_SendRequestImpl(DmaRequest* req, u32 ram, u32 vrom, u32 size, u32 unk, OSMesgQueue* queue, OSMesg msg) {
+s32 DmaMgr_SendRequestImpl(DmaRequest* req, uintptr_t ram, uintptr_t vrom, u32 size, u32 unk, OSMesgQueue* queue, OSMesg msg) {
     static s32 sDmaMgrQueueFullLogged = 0;
 
     if ((1 && (ram == 0)) || (osMemSize < ram + size + 0x80000000) || (vrom & 1) || (vrom > 0x4000000) || (size == 0) ||
@@ -363,7 +363,7 @@ s32 DmaMgr_SendRequestImpl(DmaRequest* req, u32 ram, u32 vrom, u32 size, u32 unk
     return 0;
 }
 
-s32 DmaMgr_SendRequest0(u32 ram, u32 vrom, u32 size) {
+s32 DmaMgr_SendRequest0(uintptr_t ram, uintptr_t vrom, u32 size) {
     DmaRequest req;
     OSMesgQueue queue;
     OSMesg msg;
@@ -384,8 +384,8 @@ void DmaMgr_Init(void) {
     s32 idx;
     DmaEntry* iter;
 
-    DmaMgr_DmaRomToRam((u32)_dmadataSegmentRomStart, (u32)_dmadataSegmentStart,
-                       (u32)(_dmadataSegmentRomEnd - _dmadataSegmentRomStart));
+    DmaMgr_DmaRomToRam((uintptr_t)_dmadataSegmentRomStart, (uintptr_t)_dmadataSegmentStart,
+                       (uintptr_t)(_dmadataSegmentRomEnd - _dmadataSegmentRomStart));
     osSyncPrintf("dma_rom_ad[]\n");
 
     sDmaMgrDataExistError = 0;
@@ -411,7 +411,7 @@ void DmaMgr_Init(void) {
         }
     }
 
-    if ((u32)_bootSegmentRomStart != gDmaDataTable[0].vromEnd) {
+    if ((uintptr_t)_bootSegmentRomStart != gDmaDataTable[0].vromEnd) {
         osSyncPrintf("_bootSegmentRomStart(%08x) != dma_rom_ad[0].rom_b(%08x)\n", _bootSegmentRomStart,
                      gDmaDataTable[0].vromEnd);
         Fault_AddHungupAndCrash("../z_std_dma.c", 1055);
@@ -423,19 +423,19 @@ void DmaMgr_Init(void) {
     osStartThread(&sDmaMgrThread);
 }
 
-s32 DmaMgr_SendRequest2(DmaRequest* req, u32 ram, u32 vrom, u32 size, u32 unk5, OSMesgQueue* queue, OSMesg msg,
+s32 DmaMgr_SendRequest2(DmaRequest* req, uintptr_t ram, uintptr_t vrom, u32 size, u32 unk5, OSMesgQueue* queue, OSMesg msg,
                         const char* file, s32 line) {
     req->filename = file;
     req->line = line;
     DmaMgr_SendRequestImpl(req, ram, vrom, size, unk5, queue, msg);
 }
 
-s32 DmaMgr_SendRequest1(void* ram0, u32 vrom, u32 size, const char* file, s32 line) {
+s32 DmaMgr_SendRequest1(void* ram0, uintptr_t vrom, u32 size, const char* file, s32 line) {
     DmaRequest req;
     s32 ret;
     OSMesgQueue queue;
     OSMesg msg;
-    u32 ram = (u32)ram0;
+    uintptr_t ram = (uintptr_t)ram0;
 
     req.filename = file;
     req.line = line;
